backend/requirements.txt
fastapi
uvicorn
spotipy
python-dotenv
httpx
itsdangerous
google-cloud-storage
gunicorn
